+++ src/qg_gen.cpp
@@ -8,12 +8,12 @@ int debug_callback(void *data, int size, char **var0, char **var1) {
     return 0;
 }
 
-int8_t size_to_districts[city_size::SIZE_COUNT] = { 1, 3, 5, 9 };
-int32_t district_weights[district_type::DISTRICT_COUNT] = { 2, 3, 4, 1, 1 };
+i8 size_to_districts[city_size::SIZE_COUNT] = { 1, 3, 5, 9 };
+i32 district_weights[district_type::DISTRICT_COUNT] = { 2, 3, 4, 1, 1 };
 
-int8_t size_to_landmarks[city_size::SIZE_COUNT] = { 5, 15, 40, 80 };
+i8 size_to_landmarks[city_size::SIZE_COUNT] = { 5, 15, 40, 80 };
 
-void case_gen_fondation(case_gen *ctx, city_size s, int32_t seed) {
+void case_gen_fondation(case_gen *ctx, city_size s, i32 seed) {
     srand(seed);
     ctx->size = s;
     ctx->num_landmarks = 0;
@@ -73,9 +73,9 @@ void case_gen_polish(case_gen *ctx) { }
 // ====================
 
 const char *district_prefix[] { "Little", "Grand", "Silver", "High", "Low", "Old", "New", "Upper", "Lower", "Greater" };
-const size_t num_district_prefix = sizeof(district_prefix) / sizeof(district_prefix[0]);
+const u64 num_district_prefix = sizeof(district_prefix) / sizeof(district_prefix[0]);
 const char *district_suffix[] { "Heights", "Park", "Hills", "Grove", "Valley", "District", "Quarter", "Gardens", "Square", "Town", "Village", "Estates", "Side", "End" };
-const size_t num_district_suffix = sizeof(district_suffix) / sizeof(district_suffix[0]);
+const u64 num_district_suffix = sizeof(district_suffix) / sizeof(district_suffix[0]);
 
 bool name_gen_validate(const std::string &name) {
     size_t f_space = name.find_first_of(' ');
@@ -89,12 +89,12 @@ bool name_gen_validate(const std::string &name) {
 }
 
 void name_gen_train(name_gen *gen, const char *file_path) {
-    size_t len = 0;
+    u64 len = 0;
     std::string data = (char *)SDL_LoadFile(file_path, &len);
 
-    size_t s = 0;
-    size_t e = data.find(',', s);
-    static size_t k = 3;
+    u64 s = 0;
+    u64 e = data.find(',', s);
+    static u64 k = 3;
 
     while (e != std::string::npos) {
         std::string name = "^^^" + data.substr(s, e-s) + "$$$";
@@ -110,12 +110,12 @@ void name_gen_train(name_gen *gen, const char *file_path) {
     }
 }
 
-void name_gen_next(name_gen *gen, size_t num, std::vector<std::string> *out) {
-    static size_t k = 3;
+void name_gen_next(name_gen *gen, u64 num, std::vector<std::string> *out) {
+    static u64 k = 3;
     std::vector<char> options;
     std::vector<int32_t> weights;
 
-    for (int i = 0; i < num; i++) {
+    for (u64 i = 0; i < num; i++) {
         std::string name = "^^^";
 
         char next = '\0';
@@ -144,7 +144,7 @@ void name_gen_next(name_gen *gen, size_t num, std::vector<std::string> *out) {
     }
 }
 
-void name_gen_district(name_gen *gen, size_t num, std::vector<std::string> *out) {
+void name_gen_district(name_gen *gen, u64 num, std::vector<std::string> *out) {
     name_gen_next(gen, num, out);
 
     //TODO: Is there a better way
